+++ packages/types/custom/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Feature: Add `snapTo` to `DrawConfiguration` for specification of layers to snap to.
 - Feature: Add `lassos` to `DrawConfiguration`. With this, `addLoading`, `removeLoading`, and `toastAction` have also been introduced to allow the feature to use other plugins via API calls, and `Lasso` itself has been introduced.
 - Feature: Add new interface `LayerChooserConfiguration` and the parameter `layerChooser` to `MapConfig` for the configurability of `@polar/plugin-layer-chooser`.
+- Feature: Add new parameter `component` to `AddressSearchConfiguration` for the configurability of the shown component.
 - Feature: Add new parameter `component` to `IconMenuConfiguration` for the configurability of the shown component.
 - Feature: Add new parameter `component` to `ZoomConfiguration` for the configurability of the shown component.
 - Feature: Add new parameter `buttonComponent` to `AttributionsConfiguration` for the configurability of the button component in `renderType` `independent`.